pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "flask>=3.1.2",
    "google-genai>=1.59.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
    "myproject",
]
